[apps] :recycle: Clean-up sign conversion warnings from the convolution kernels

+++ apps/common/kernel/convolution.h
@@ -27,29 +27,30 @@ void conv2d_parallel(int32_t const *__restrict__ in, uint32_t in_x,
                      uint32_t k_x, uint32_t k_y,
                      int32_t volatile *__restrict__ out, uint32_t id,
                      uint32_t numThreads) {
-  int boundary_x = k_x / 2;
-  int boundary_y = k_y / 2;
+  int boundary_x = (int)(k_x / 2);
+  int boundary_y = (int)(k_y / 2);
   // Now we only care about valid entries
-  while (id < boundary_x) {
+  while (id < (unsigned int)boundary_x) {
     id += numThreads;
   }
   int32_t sum;
   uint32_t weight = 0;
-  for (int i = 0; i < k_x * k_y; ++i) {
+  for (unsigned int i = 0; i < k_x * k_y; ++i) {
     weight += k[i];
   }
   // TODO implement boundary halo
   // Start at the boundary_x
-  for (int i = id; i < in_x - boundary_x; i += numThreads) {
-    for (int j = boundary_y; j < in_y - boundary_y; j++) {
+  for (int i = (int)id; i < (int)in_x - boundary_x; i += (int)numThreads) {
+    for (int j = boundary_y; j < (int)in_y - boundary_y; j++) {
       sum = 0;
-      for (int m = -boundary_y; m < (int)(k_y - boundary_y); m++) {
-        for (int n = -boundary_x; n < (int)(k_x - boundary_x); n++) {
-          sum += in[(j + m) * in_x + (i + n)] *
-                 k[(m + boundary_y) * k_x + (n + boundary_x)];
+      for (int m = -boundary_y; m < (int)k_y - boundary_y; m++) {
+        for (int n = -boundary_x; n < (int)k_x - boundary_x; n++) {
+          sum += in[(unsigned int)(j + m) * in_x + (unsigned int)(i + n)] *
+                 (int)k[(unsigned int)(m + boundary_y) * k_x +
+                        (unsigned int)(n + boundary_x)];
         }
       }
-      out[j * in_x + i] = sum / weight;
+      out[(unsigned int)j * in_x + (unsigned int)i] = sum / (int)weight;
     }
   }
 }
@@ -59,24 +60,26 @@ void conv2d_shifted_parallel(int32_t const *__restrict__ in, uint32_t in_x,
                              uint32_t k_x, uint32_t k_y,
                              int32_t volatile *__restrict__ out, uint32_t id,
                              uint32_t numThreads) {
-  uint32_t boundary_x = k_x / 2;
-  uint32_t boundary_y = k_y / 2;
+  int boundary_x = (int)(k_x / 2);
+  int boundary_y = (int)(k_y / 2);
   int32_t sum;
   uint32_t weight = 0;
-  for (int i = 0; i < k_x * k_y; ++i) {
+  for (unsigned int i = 0; i < k_x * k_y; ++i) {
     weight += k[i];
   }
   // TODO implement boundary halo
   // Now we only care about valid entries
-  for (uint32_t i = id; i < in_x - (2 * boundary_x); i += numThreads) {
-    for (uint32_t j = 0; j < in_y - (2 * boundary_y); j++) {
+  for (unsigned int i = id; i < in_x - (unsigned int)(2 * boundary_x);
+       i += numThreads) {
+    for (unsigned int j = 0; j < in_y - (unsigned int)(2 * boundary_y); j++) {
       sum = 0;
-      for (uint32_t m = 0; m < k_y; m++) {
-        for (uint32_t n = 0; n < k_x; n++) {
-          sum += in[(j + m) * in_x + (i + n)] * k[m * k_x + n];
+      for (unsigned int m = 0; m < k_y; m++) {
+        for (unsigned int n = 0; n < k_x; n++) {
+          sum += in[(j + m) * in_x + (i + n)] * (int)k[m * k_x + n];
         }
       }
-      out[(j + boundary_y) * in_x + (i + boundary_x)] = sum / weight;
+      out[(j + (unsigned int)boundary_y) * in_x +
+          (i + (unsigned int)boundary_x)] = sum / (int)weight;
     }
   }
 }
@@ -87,7 +90,7 @@ void conv2d_3x3_unrolled_parallel(int32_t const *__restrict__ in, uint32_t in_x,
                                   uint32_t id, uint32_t numThreads) {
   int32_t sum;
   uint32_t weight = 0;
-  for (int i = 0; i < 9; ++i) {
+  for (unsigned int i = 0; i < 9; ++i) {
     weight += k[i];
   }
   // TODO implement boundary halo
@@ -109,16 +112,16 @@ void conv2d_3x3_unrolled_parallel(int32_t const *__restrict__ in, uint32_t in_x,
   for (uint32_t i = start; i < end; ++i) {
     for (uint32_t j = 1; j < in_y - 1; j++) {
       sum = 0;
-      sum += in[(j - 1) * in_x + (i - 1)] * k[0];
-      sum += in[(j - 1) * in_x + (i + 0)] * k[1];
-      sum += in[(j - 1) * in_x + (i + 1)] * k[2];
-      sum += in[(j + 0) * in_x + (i - 1)] * k[3];
-      sum += in[(j + 0) * in_x + (i + 0)] * k[4];
-      sum += in[(j + 0) * in_x + (i + 1)] * k[5];
-      sum += in[(j + 1) * in_x + (i - 1)] * k[6];
-      sum += in[(j + 1) * in_x + (i + 0)] * k[7];
-      sum += in[(j + 1) * in_x + (i + 1)] * k[8];
-      out[j * in_x + i] = sum / weight;
+      sum += in[(j - 1) * in_x + (i - 1)] * (int)k[0];
+      sum += in[(j - 1) * in_x + (i + 0)] * (int)k[1];
+      sum += in[(j - 1) * in_x + (i + 1)] * (int)k[2];
+      sum += in[(j + 0) * in_x + (i - 1)] * (int)k[3];
+      sum += in[(j + 0) * in_x + (i + 0)] * (int)k[4];
+      sum += in[(j + 0) * in_x + (i + 1)] * (int)k[5];
+      sum += in[(j + 1) * in_x + (i - 1)] * (int)k[6];
+      sum += in[(j + 1) * in_x + (i + 0)] * (int)k[7];
+      sum += in[(j + 1) * in_x + (i + 1)] * (int)k[8];
+      out[j * in_x + i] = sum / (int)weight;
     }
   }
 }
@@ -135,19 +138,19 @@ void conv2d_3x3_shifted_unrolled_parallel(int32_t const *__restrict__ in,
   }
   // TODO implement boundary halo
   // Now we only care about valid entries
-  for (int i = id; i < in_x - 2; i += numThreads) {
-    for (int j = 0; j < in_y - 2; j++) {
+  for (unsigned int i = id; i < in_x - 2; i += numThreads) {
+    for (unsigned int j = 0; j < in_y - 2; j++) {
       sum = 0;
-      sum += in[(j + 0) * in_x + (i + 0)] * k[0];
-      sum += in[(j + 0) * in_x + (i + 1)] * k[1];
-      sum += in[(j + 0) * in_x + (i + 2)] * k[2];
-      sum += in[(j + 1) * in_x + (i + 0)] * k[3];
-      sum += in[(j + 1) * in_x + (i + 1)] * k[4];
-      sum += in[(j + 1) * in_x + (i + 2)] * k[5];
-      sum += in[(j + 2) * in_x + (i + 0)] * k[6];
-      sum += in[(j + 2) * in_x + (i + 1)] * k[7];
-      sum += in[(j + 2) * in_x + (i + 2)] * k[8];
-      out[(j + 1) * in_x + (i + 1)] = sum / weight;
+      sum += in[(j + 0) * in_x + (i + 0)] * (int)k[0];
+      sum += in[(j + 0) * in_x + (i + 1)] * (int)k[1];
+      sum += in[(j + 0) * in_x + (i + 2)] * (int)k[2];
+      sum += in[(j + 1) * in_x + (i + 0)] * (int)k[3];
+      sum += in[(j + 1) * in_x + (i + 1)] * (int)k[4];
+      sum += in[(j + 1) * in_x + (i + 2)] * (int)k[5];
+      sum += in[(j + 2) * in_x + (i + 0)] * (int)k[6];
+      sum += in[(j + 2) * in_x + (i + 1)] * (int)k[7];
+      sum += in[(j + 2) * in_x + (i + 2)] * (int)k[8];
+      out[(j + 1) * in_x + (i + 1)] = sum / (int)weight;
     }
   }
 }
@@ -158,15 +161,15 @@ void init_conv2d_image(volatile int32_t *img, uint32_t img_x, uint32_t img_y,
                        uint32_t id, uint32_t numThreads) {
   // Parallelize over rows
   if (img_y > img_x) {
-    for (int i = id; i < img_y; i += numThreads) {
-      for (int j = 0; j < img_x; ++j) {
-        img[i * img_x + j] = (i % 16) + (j % 4);
+    for (int i = (int)id; i < (int)img_y; i += (int)numThreads) {
+      for (int j = 0; j < (int)img_x; ++j) {
+        img[(unsigned int)i * img_x + (unsigned int)j] = (i % 16) + (j % 4);
       }
     }
   } else {
-    for (int j = id; j < img_x; j += numThreads) {
-      for (int i = 0; i < img_y; ++i) {
-        img[i * img_x + j] = (i % 16) + (j % 4);
+    for (int j = (int)id; j < (int)img_x; j += (int)numThreads) {
+      for (int i = 0; i < (int)img_y; ++i) {
+        img[(unsigned int)i * img_x + (unsigned int)j] = (i % 16) + (j % 4);
       }
     }
   }
@@ -177,15 +180,15 @@ void zero_conv2d_image(volatile int32_t *img, uint32_t img_x, uint32_t img_y,
                        uint32_t id, uint32_t numThreads) {
   // Parallelize over rows
   if (img_y > img_x) {
-    for (int i = id; i < img_y; i += numThreads) {
-      for (int j = 0; j < img_x; ++j) {
-        img[i * img_x + j] = 0;
+    for (int i = (int)id; i < (int)img_y; i += (int)numThreads) {
+      for (int j = 0; j < (int)img_x; ++j) {
+        img[(unsigned int)i * img_x + (unsigned int)j] = 0;
       }
     }
   } else {
-    for (int j = id; j < img_x; j += numThreads) {
-      for (int i = 0; i < img_y; ++i) {
-        img[i * img_x + j] = 0;
+    for (int j = (int)id; j < (int)img_x; j += (int)numThreads) {
+      for (int i = 0; i < (int)img_y; ++i) {
+        img[(unsigned int)i * img_x + (unsigned int)j] = 0;
       }
     }
   }
@@ -197,18 +200,18 @@ extern uint32_t barrier_init;
 int verify_conv2d_image(volatile int32_t *img, uint32_t img_x, uint32_t img_y,
                         uint32_t id, uint32_t numThreads) {
   // Parallelize over rows
-  for (int i = id + 1; i < img_y - 1; i += numThreads) {
-    int32_t y = i % 16;
+  for (int i = (int)id + 1; i < (int)img_y - 1; i += (int)numThreads) {
+    int y = i % 16;
     if (i % 16 == 0)
       y = 4;
     if (i % 16 == 15)
       y = 11;
-    for (int32_t j = 1; j < img_x - 1; ++j) {
-      int32_t x = ((j % 4) / 2) + 1;
-      if (img[i * img_x + j] != x + y) {
-        return (i + j) == 0 ? -1 : i * img_x + j;
+    for (int j = 1; j < (int)img_x - 1; ++j) {
+      int x = ((j % 4) / 2) + 1;
+      if ((int)img[i * (int)img_x + j] != x + y) {
+        return (i + j) == 0 ? -1 : i * (int)img_x + j;
       }
-      img[i * img_x + j] = 0;
+      img[i * (int)img_x + j] = 0;
     }
   }
   return 0;
